Restructure OTG-FS driver
