+++ esp-hal/CHANGELOG.md
@@ -147,6 +147,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `esp_hal::efuse::chip_revision` has been marked stable (#5287)
 - `esp_hal::efuse::chip_revision` now returns `ChipRevision` (#5287)
 - DMA buffers can now be created using empty buffer slices (#5266)
+- The embassy-usb device drivers has been moved from `esp_hal::otg_fs::asynch` to `esp_hal::otg_fs::device`. (#5283)
 
 ### Fixed
